+++ src/gameObject.hpp
@@ -15,7 +15,7 @@ namespace iso {
 
         public:
             std::string    object_name;         // Object's name indicates the object type -> "Building" or "Region" or "Tile".
-            core::Vector3f object_position;     // Object's position in the game world.
+            core::Vector3i object_position;     // Object's position in the game world.
             core::Vector2i object_size;         // Object's size, equivalent of the texture size.
             std::string    object_texture_name; // Object's texture name.
             core::Colour   object_colour;
@@ -23,17 +23,17 @@ namespace iso {
         public:
             GameObject();
             GameObject(const GameObject& object);
-            GameObject(core::Vector3f position, core::Vector3f relative_position, core::Vector2i size, const std::string& texture_name);
-            GameObject(core::Vector3f position, core::Vector3f relative_position, core::Vector2i size, const std::string& texture_name, const std::string& object_name);
+            GameObject(core::Vector3i position, core::Vector3i relative_position, core::Vector2i size, const std::string& texture_name);
+            GameObject(core::Vector3i position, core::Vector3i relative_position, core::Vector2i size, const std::string& texture_name, const std::string& object_name);
             ~GameObject();
             
             bool exists() const;
             bool hasName() const;
-            bool contains(core::Vector2f point) const;
+            bool contains(core::Vector2i point) const;
 
             const std::string& getName()        const;
-            core::Vector3f     getPosition()    const;
-            core::Vector2f     getPosition2D()  const;
+            core::Vector3i     getPosition()    const;
+            core::Vector2i     getPosition2D()  const;
             core::Vector2i     getSize()        const;
             const std::string& getTextureName() const;
             core::Colour       getColour()      const;
